### Does this PR have any customer impact?
No

### Type (Feature, Refactoring, Bugfix, DevOps, Testing, Perf, etc)
Refactor

### Does it involve schema level changes? (Table, Column, Index, UDF, etc level changes)
No

### Are you introducing any new config? If yes, do you have tests with and without them being set?
No
### Release Note (Refer [Template](../docs/release_notes/pgmongo_template.md))

### Description
Split up GUCs by intent:
- Testing GUCs - only for backend or test infra
- System GUCs - configuration of the system as a whole
- Background GUCs - configuration for background jobs
- Feature flag GUCs - GUCs that are transient and we should track enabling and removing them

----
#### AI description  (iteration 1)
#### PR Classification
Code cleanup

#### PR Summary
This pull request refactors the codebase by splitting up GUC (Grand Unified Configuration) files based on their intent, improving modularity and maintainability.
- Refactored `documentdb_init.c` to delegate GUC initialization to specific functions.
- Added new files `feature_flag_configs.c`, `system_configs.c`, `background_job_configs.c`, and `testing_configs.c` for modular GUC initialization.
- Updated `config_initialization.h` to declare new initialization functions.
- Adjusted related configurations in `documentdb_distributed_init.c` and test files to align with the new structure.

+++ pg_documentdb/src/opclass/rum_exclusion.c
@@ -60,7 +60,7 @@
 /* --------------------------------------------------------- */
 
 extern bool ForceEnableNewUniqueOpClass;
-extern uint64_t DefaultUniqueIndexKeyhashOverride;
+extern int DefaultUniqueIndexKeyhashOverride;
 
 static pgbson * GetShardKeyAndDocument(HeapTupleHeader input, int64_t *shardKey);
 static IndexTraverseOption GetExclusionIndexTraverseOption(void *contextOptions,
@@ -578,7 +578,7 @@ ExtractUniqueShardTermsFromInput(pgbson *input, int32_t *nentries, Pointer **ext
 			 * This is only used for testing and should not be set in production.
 			 */
 			*lastBytes = DefaultUniqueIndexKeyhashOverride > 0 ?
-						 DefaultUniqueIndexKeyhashOverride :
+						 (uint64_t) DefaultUniqueIndexKeyhashOverride :
 						 HashBsonValueComparableExtended(indexTerm, keyhash);
 
 			if (index > numTerms)
